# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "Holiday": {
        "properties": {
          "calendar_id": {
            "format": "uuid",
            "title": "Calendar Id",
            "type": "string"
          },
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "estimated": {
            "title": "Estimated",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "is_non_working": {
            "title": "Is Non Working",
            "type": "boolean"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "source": {
            "enum": [
              "bundled",
              "user"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "id",
          "calendar_id",
          "date",
          "name",
          "is_non_working",
          "source",
          "estimated"
        ],
        "title": "Holiday",
        "type": "object"
      },
      "HolidayCalendar": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "colour": {
            "enum": [
              "slate",
              "red",
              "orange",
              "amber",
              "lime",
              "green",
              "teal",
              "cyan",
              "blue",
              "indigo",
              "violet",
              "pink"
            ],
            "title": "Colour",
            "type": "string"
          },
          "enabled": {
            "title": "Enabled",
            "type": "boolean"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "code",
          "name",
          "enabled",
          "colour"
        ],
        "title": "HolidayCalendar",
        "type": "object"
      },
      "HolidayCalendarUpdate": {
        "properties": {
          "colour": {
            "anyOf": [
              {
                "enum": [
                  "slate",
                  "red",
                  "orange",
                  "amber",
                  "lime",
                  "green",
                  "teal",
                  "cyan",
                  "blue",
                  "indigo",
                  "violet",
                  "pink"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Colour"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          }
        },
        "title": "HolidayCalendarUpdate",
        "type": "object"
      },
      "HolidayCreate": {
        "properties": {
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "is_non_working": {
            "default": true,
            "title": "Is Non Working",
            "type": "boolean"
          },
          "name": {
            "maxLength": 100,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "date",
          "name"
        ],
        "title": "HolidayCreate",
        "type": "object"
      },
      "HolidayUpdate": {
        "properties": {
          "date": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date"
          },
          "is_non_working": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Non Working"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 100,
                "minLength": 1,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "HolidayUpdate",
        "type": "object"
      },
      "LeaveBalance": {
        "properties": {
          "allowance_days": {
            "title": "Allowance Days",
            "type": "number"
          },
          "bookings": {
            "items": {
              "$ref": "#/components/schemas/LeaveBooking"
            },
            "title": "Bookings",
            "type": "array"
          },
          "carried_over_days": {
            "title": "Carried Over Days",
            "type": "number"
          },
          "planned": {
            "title": "Planned",
            "type": "number"
          },
          "remaining": {
            "title": "Remaining",
            "type": "number"
          },
          "used": {
            "title": "Used",
            "type": "number"
          },
          "year": {
            "title": "Year",
            "type": "integer"
          }
        },
        "required": [
          "year",
          "allowance_days",
          "carried_over_days",
          "used",
          "planned",
          "remaining",
          "bookings"
        ],
        "title": "LeaveBalance",
        "type": "object"
      },
...
      "LeaveImpact": {
        "properties": {
          "days": {
            "title": "Days",
            "type": "number"
          },
          "exceeds": {
            "title": "Exceeds",
            "type": "boolean"
          },
          "remaining_after": {
            "title": "Remaining After",
            "type": "number"
          },
          "remaining_before": {
            "title": "Remaining Before",
            "type": "number"
          },
          "year": {
            "title": "Year",
            "type": "integer"
          }
        },
        "required": [
          "year",
          "days",
          "remaining_before",
          "remaining_after",
          "exceeds"
        ],
        "title": "LeaveImpact",
        "type": "object"
      },
      "LeavePolicy": {
        "properties": {
          "allowance_days": {
            "title": "Allowance Days",
            "type": "number"
          },
          "carried_over_days": {
            "title": "Carried Over Days",
            "type": "number"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "year": {
            "title": "Year",
            "type": "integer"
          }
        },
        "required": [
          "year",
          "allowance_days",
          "carried_over_days",
          "version"
        ],
        "title": "LeavePolicy",
        "type": "object"
      },
      "LeavePolicyUpdate": {
        "properties": {
          "allowance_days": {
            "anyOf": [
              {
                "maximum": 999.9,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,3}|(?=[\\d.]{1,5}0*$)\\d{0,3}\\.\\d{0,1}0*$)",
                "type": "string"
              }
            ],
            "title": "Allowance Days"
          },
          "carried_over_days": {
            "anyOf": [
              {
                "maximum": 999.9,
                "minimum": 0.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*(?:\\d{0,3}|(?=[\\d.]{1,5}0*$)\\d{0,3}\\.\\d{0,1}0*$)",
                "type": "string"
              }
            ],
            "title": "Carried Over Days"
          },
          "version": {
            "anyOf": [
              {
                "minimum": 1.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Version"
          }
        },
        "required": [
          "allowance_days",
          "carried_over_days"
        ],
        "title": "LeavePolicyUpdate",
        "type": "object"
      },
      "LeavePreviewRequest": {
        "properties": {
          "category_id": {
            "format": "uuid",
            "title": "Category Id",
            "type": "string"
          },
          "end_date": {
            "format": "date",
            "title": "End Date",
            "type": "string"
          },
          "exclude_event_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exclude Event Id"
          },
          "repeat": {
            "default": "none",
            "enum": [
              "none",
              "monthly",
              "yearly"
            ],
            "title": "Repeat",
            "type": "string"
          },
          "repeat_until": {
            "anyOf": [
              {
                "format": "date",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Repeat Until"
          },
          "start_date": {
            "format": "date",
            "title": "Start Date",
            "type": "string"
          }
        },
        "required": [
          "start_date",
          "end_date",
          "category_id"
        ],
        "title": "LeavePreviewRequest",
        "type": "object"
      },
...
      "Problem": {
        "description": "RFC 9457 problem details, served as application/problem+json.",
        "properties": {
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          },
          "errors": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/ProblemError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errors"
          },
          "instance": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instance"
          },
          "status": {
            "title": "Status",
            "type": "integer"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "type": {
            "default": "about:blank",
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "title",
          "status"
        ],
        "title": "Problem",
        "type": "object"
      },
      "ProblemError": {
        "description": "One validation error inside a problem document.",
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Loc",
            "type": "array"
          },
          "msg": {
            "title": "Msg",
            "type": "string"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ProblemError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Self-hosted personal planner.",
    "title": "Hoje API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/holiday-calendars": {
      "get": {
        "operationId": "holiday_calendars_list",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/HolidayCalendar"
                  },
                  "title": "Response Holiday Calendars List",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "CSRF or Origin check failed, or forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not implemented yet"
          }
        },
        "summary": "List holiday calendars",
        "tags": [
          "holidays"
        ]
      }
    },
    "/api/v1/holiday-calendars/{calendar_id}": {
      "patch": {
        "operationId": "holiday_calendars_update",
        "parameters": [
          {
            "in": "path",
            "name": "calendar_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Calendar Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HolidayCalendarUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HolidayCalendar"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "CSRF or Origin check failed, or forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not implemented yet"
          }
        },
        "summary": "Enable or recolour a calendar",
        "tags": [
          "holidays"
        ]
      }
    },
    "/api/v1/holiday-calendars/{calendar_id}/holidays": {
      "post": {
        "operationId": "holidays_create",
        "parameters": [
          {
            "in": "path",
            "name": "calendar_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Calendar Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HolidayCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Holiday"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "CSRF or Origin check failed, or forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not implemented yet"
          }
        },
        "summary": "Add a custom holiday",
        "tags": [
          "holidays"
        ]
      }
    },
    "/api/v1/holiday-calendars/{calendar_id}/reset": {
      "post": {
        "operationId": "holiday_calendars_reset",
        "parameters": [
          {
            "in": "path",
            "name": "calendar_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Calendar Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HolidayCalendar"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "CSRF or Origin check failed, or forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not implemented yet"
          }
        },
        "summary": "Reset a calendar to the bundled holidays",
        "tags": [
          "holidays"
        ]
      }
    },
    "/api/v1/holidays": {
      "get": {
        "operationId": "holidays_list",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": true,
            "schema": {
              "format": "date",
              "title": "From",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": true,
            "schema": {
              "format": "date",
              "title": "To",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Holiday"
                  },
                  "title": "Response Holidays List",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "CSRF or Origin check failed, or forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not implemented yet"
          }
        },
        "summary": "List holidays of enabled calendars",
        "tags": [
          "holidays"
        ]
      }
    },
    "/api/v1/holidays/{holiday_id}": {
      "delete": {
        "operationId": "holidays_delete",
        "parameters": [
          {
            "in": "path",
            "name": "holiday_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Holiday Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "CSRF or Origin check failed, or forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not implemented yet"
          }
        },
        "summary": "Delete a holiday",
        "tags": [
          "holidays"
        ]
      },
      "patch": {
        "operationId": "holidays_update",
        "parameters": [
          {
            "in": "path",
            "name": "holiday_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Holiday Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HolidayUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Holiday"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "CSRF or Origin check failed, or forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not implemented yet"
          }
        },
        "summary": "Edit a holiday",
        "tags": [
          "holidays"
        ]
      }
    },
    "/api/v1/leave/balance": {
      "get": {
        "operationId": "leave_balance",
        "parameters": [
          {
            "in": "query",
            "name": "year",
            "required": true,
            "schema": {
              "maximum": 2100,
              "minimum": 2000,
              "title": "Year",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeaveBalance"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "CSRF or Origin check failed, or forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not implemented yet"
          }
        },
        "summary": "Leave balance for a year",
        "tags": [
          "leave"
        ]
      }
    },
    "/api/v1/leave/policies/{year}": {
      "put": {
        "operationId": "leave_policies_upsert",
        "parameters": [
          {
            "in": "path",
            "name": "year",
            "required": true,
            "schema": {
              "maximum": 2100,
              "minimum": 2000,
              "title": "Year",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeavePolicyUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeavePolicy"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "CSRF or Origin check failed, or forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not implemented yet"
          }
        },
        "summary": "Set the yearly leave policy",
        "tags": [
          "leave"
        ]
      }
    },
    "/api/v1/leave/preview": {
      "post": {
        "operationId": "leave_preview",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeavePreviewRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/LeaveImpact"
                  },
                  "title": "Response Leave Preview",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not authenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "CSRF or Origin check failed, or forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Validation error"
          },
          "501": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Problem"
                }
              }
            },
            "description": "Not implemented yet"
          }
        },
        "summary": "Preview leave impact",
        "tags": [
          "leave"
        ]
      }
    },
...
  }
}